feat(girth): moving infra back to ihe stack
Refs: #1667
