+++ packages/infra/lib/api-stack.ts
@@ -347,46 +347,6 @@ export class APIStack extends Stack {
       sentryDsn: props.config.lambdasSentryDSN,
     });
 
-    const girthPatientDiscoveryLambda = this.setupGirthPatientDiscoveryLambda({
-      lambdaLayers,
-      vpc: this.vpc,
-      secrets,
-      cqOrgCertificate: props.config.carequality.secretNames.CQ_ORG_CERTIFICATE,
-      cqOrgPrivateKey: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY,
-      cqOrgCertificateIntermediate:
-        props.config.carequality.secretNames.CQ_ORG_CERTIFICATE_INTERMEDIATE,
-      cqOrgPrivateKeyPassword: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY_PASSWORD,
-      apiURL: props.config.loadBalancerDnsName,
-      envType: props.config.environmentType,
-      sentryDsn: props.config.lambdasSentryDSN,
-    });
-    const girthDocumentQueryLambda = this.setupGirthDocumentQueryLambda({
-      lambdaLayers,
-      vpc: this.vpc,
-      secrets,
-      cqOrgCertificate: props.config.carequality.secretNames.CQ_ORG_CERTIFICATE,
-      cqOrgPrivateKey: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY,
-      cqOrgCertificateIntermediate:
-        props.config.carequality.secretNames.CQ_ORG_CERTIFICATE_INTERMEDIATE,
-      cqOrgPrivateKeyPassword: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY_PASSWORD,
-      apiURL: props.config.loadBalancerDnsName,
-      envType: props.config.environmentType,
-      sentryDsn: props.config.lambdasSentryDSN,
-    });
-    const girthDocumentRetrievalLambda = this.setupGirthDocumentRetrievalLambda({
-      lambdaLayers,
-      vpc: this.vpc,
-      secrets,
-      cqOrgCertificate: props.config.carequality.secretNames.CQ_ORG_CERTIFICATE,
-      cqOrgPrivateKey: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY,
-      cqOrgCertificateIntermediate:
-        props.config.carequality.secretNames.CQ_ORG_CERTIFICATE_INTERMEDIATE,
-      cqOrgPrivateKeyPassword: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY_PASSWORD,
-      apiURL: props.config.loadBalancerDnsName,
-      envType: props.config.environmentType,
-      sentryDsn: props.config.lambdasSentryDSN,
-    });
-
     const outboundPatientDiscoveryLambda = this.setupOutboundPatientDiscovery({
       lambdaLayers,
       vpc: this.vpc,
@@ -477,9 +437,6 @@ export class APIStack extends Stack {
       outboundPatientDiscoveryLambda,
       outboundDocumentQueryLambda,
       outboundDocumentRetrievalLambda,
-      girthPatientDiscoveryLambda,
-      girthDocumentQueryLambda,
-      girthDocumentRetrievalLambda,
       medicalDocumentsUploadBucket,
       fhirToMedicalRecordLambda,
       searchIngestionQueue: ccdaSearchQueue,
@@ -1376,225 +1333,6 @@ export class APIStack extends Stack {
     dbCredsSecret.grantRead(outboundDocumentRetrievalLambda);
 
     return outboundDocumentRetrievalLambda;
-  }
-
-  private setupGirthPatientDiscoveryLambda(ownProps: {
-    lambdaLayers: LambdaLayers;
-    vpc: ec2.IVpc;
-    secrets: Secrets;
-    cqOrgCertificate: string;
-    cqOrgPrivateKey: string;
-    cqOrgPrivateKeyPassword: string;
-    cqOrgCertificateIntermediate: string;
-    apiURL: string;
-    envType: EnvType;
-    sentryDsn: string | undefined;
-  }): Lambda {
-    const {
-      lambdaLayers,
-      vpc,
-      secrets,
-      cqOrgCertificate,
-      cqOrgPrivateKey,
-      cqOrgPrivateKeyPassword,
-      cqOrgCertificateIntermediate,
-      apiURL,
-      envType,
-      sentryDsn,
-    } = ownProps;
-
-    const patientDiscoveryLambda = createLambda({
-      stack: this,
-      name: "GirthOutboundPatientDiscovery",
-      runtime: lambda.Runtime.NODEJS_18_X,
-      entry: "girth-outbound-patient-discovery",
-      envType: envType,
-      envVars: {
-        CQ_ORG_PRIVATE_KEY: cqOrgPrivateKey,
-        CQ_ORG_CERTIFICATE: cqOrgCertificate,
-        CQ_ORG_CERTIFICATE_INTERMEDIATE: cqOrgCertificateIntermediate,
-        CQ_ORG_PRIVATE_KEY_PASSWORD: cqOrgPrivateKeyPassword,
-        API_URL: apiURL,
-        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
-      },
-      layers: [lambdaLayers.shared],
-      memory: 512,
-      timeout: Duration.minutes(5),
-      vpc,
-    });
-
-    // granting secrets read access to lambda
-    const cqOrgCertificateKey = "CQ_ORG_CERTIFICATE";
-    if (!secrets[cqOrgCertificateKey]) {
-      throw new Error(`${cqOrgCertificateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateKey]?.grantRead(patientDiscoveryLambda);
-
-    const cqOrgCertificateIntermediateKey = "CQ_ORG_CERTIFICATE_INTERMEDIATE";
-    if (!secrets[cqOrgCertificateIntermediateKey]) {
-      throw new Error(`${cqOrgCertificateIntermediateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateIntermediateKey]?.grantRead(patientDiscoveryLambda);
-
-    const cqOrgPrivateKeyKey = "CQ_ORG_PRIVATE_KEY";
-    if (!secrets[cqOrgPrivateKeyKey]) {
-      throw new Error(`${cqOrgPrivateKeyKey} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyKey]?.grantRead(patientDiscoveryLambda);
-
-    const cqOrgPrivateKeyPasswordKey = "CQ_ORG_PRIVATE_KEY_PASSWORD";
-    if (!secrets[cqOrgPrivateKeyPasswordKey]) {
-      throw new Error(`${cqOrgPrivateKeyPassword} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyPasswordKey]?.grantRead(patientDiscoveryLambda);
-
-    return patientDiscoveryLambda;
-  }
-
-  private setupGirthDocumentQueryLambda(ownProps: {
-    lambdaLayers: LambdaLayers;
-    vpc: ec2.IVpc;
-    secrets: Secrets;
-    cqOrgCertificate: string;
-    cqOrgPrivateKey: string;
-    cqOrgPrivateKeyPassword: string;
-    cqOrgCertificateIntermediate: string;
-    apiURL: string;
-    envType: EnvType;
-    sentryDsn: string | undefined;
-  }): Lambda {
-    const {
-      lambdaLayers,
-      vpc,
-      secrets,
-      cqOrgCertificate,
-      cqOrgPrivateKey,
-      cqOrgPrivateKeyPassword,
-      cqOrgCertificateIntermediate,
-      apiURL,
-      envType,
-      sentryDsn,
-    } = ownProps;
-
-    const documentQueryLambda = createLambda({
-      stack: this,
-      name: "GirthOutboundDocumentQuery",
-      runtime: lambda.Runtime.NODEJS_18_X,
-      entry: "girth-outbound-document-query",
-      envType: envType,
-      envVars: {
-        CQ_ORG_PRIVATE_KEY: cqOrgPrivateKey,
-        CQ_ORG_CERTIFICATE: cqOrgCertificate,
-        CQ_ORG_CERTIFICATE_INTERMEDIATE: cqOrgCertificateIntermediate,
-        CQ_ORG_PRIVATE_KEY_PASSWORD: cqOrgPrivateKeyPassword,
-        API_URL: apiURL,
-        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
-      },
-      layers: [lambdaLayers.shared],
-      memory: 512,
-      timeout: Duration.minutes(5),
-      vpc,
-    });
-
-    // granting secrets read access to lambda
-    const cqOrgCertificateKey = "CQ_ORG_CERTIFICATE";
-    if (!secrets[cqOrgCertificateKey]) {
-      throw new Error(`${cqOrgCertificateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateKey]?.grantRead(documentQueryLambda);
-
-    const cqOrgCertificateIntermediateKey = "CQ_ORG_CERTIFICATE_INTERMEDIATE";
-    if (!secrets[cqOrgCertificateIntermediateKey]) {
-      throw new Error(`${cqOrgCertificateIntermediateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateIntermediateKey]?.grantRead(documentQueryLambda);
-
-    const cqOrgPrivateKeyKey = "CQ_ORG_PRIVATE_KEY";
-    if (!secrets[cqOrgPrivateKeyKey]) {
-      throw new Error(`${cqOrgPrivateKeyKey} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyKey]?.grantRead(documentQueryLambda);
-
-    const cqOrgPrivateKeyPasswordKey = "CQ_ORG_PRIVATE_KEY_PASSWORD";
-    if (!secrets[cqOrgPrivateKeyPasswordKey]) {
-      throw new Error(`${cqOrgPrivateKeyPassword} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyPasswordKey]?.grantRead(documentQueryLambda);
-
-    return documentQueryLambda;
-  }
-
-  private setupGirthDocumentRetrievalLambda(ownProps: {
-    lambdaLayers: LambdaLayers;
-    vpc: ec2.IVpc;
-    secrets: Secrets;
-    cqOrgCertificate: string;
-    cqOrgPrivateKey: string;
-    cqOrgPrivateKeyPassword: string;
-    cqOrgCertificateIntermediate: string;
-    apiURL: string;
-    envType: EnvType;
-    sentryDsn: string | undefined;
-  }): Lambda {
-    const {
-      lambdaLayers,
-      vpc,
-      secrets,
-      cqOrgCertificate,
-      cqOrgPrivateKey,
-      cqOrgPrivateKeyPassword,
-      cqOrgCertificateIntermediate,
-      apiURL,
-      envType,
-      sentryDsn,
-    } = ownProps;
-
-    const documentRetrievalLambda = createLambda({
-      stack: this,
-      name: "GirthOutboundDocumentRetrieval",
-      runtime: lambda.Runtime.NODEJS_18_X,
-      entry: "girth-outbound-document-retrieval",
-      envType: envType,
-      envVars: {
-        CQ_ORG_PRIVATE_KEY: cqOrgPrivateKey,
-        CQ_ORG_CERTIFICATE: cqOrgCertificate,
-        CQ_ORG_CERTIFICATE_INTERMEDIATE: cqOrgCertificateIntermediate,
-        CQ_ORG_PRIVATE_KEY_PASSWORD: cqOrgPrivateKeyPassword,
-        API_URL: apiURL,
-        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
-      },
-      layers: [lambdaLayers.shared],
-      memory: 512,
-      timeout: Duration.minutes(5),
-      vpc,
-    });
-
-    // granting secrets read access to lambda
-    const cqOrgCertificateKey = "CQ_ORG_CERTIFICATE";
-    if (!secrets[cqOrgCertificateKey]) {
-      throw new Error(`${cqOrgCertificateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateKey]?.grantRead(documentRetrievalLambda);
-
-    const cqOrgCertificateIntermediateKey = "CQ_ORG_CERTIFICATE_INTERMEDIATE";
-    if (!secrets[cqOrgCertificateIntermediateKey]) {
-      throw new Error(`${cqOrgCertificateIntermediateKey} is not defined in config`);
-    }
-    secrets[cqOrgCertificateIntermediateKey]?.grantRead(documentRetrievalLambda);
-
-    const cqOrgPrivateKeyKey = "CQ_ORG_PRIVATE_KEY";
-    if (!secrets[cqOrgPrivateKeyKey]) {
-      throw new Error(`${cqOrgPrivateKeyKey} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyKey]?.grantRead(documentRetrievalLambda);
-
-    const cqOrgPrivateKeyPasswordKey = "CQ_ORG_PRIVATE_KEY_PASSWORD";
-    if (!secrets[cqOrgPrivateKeyPasswordKey]) {
-      throw new Error(`${cqOrgPrivateKeyPassword} is not defined in config`);
-    }
-    secrets[cqOrgPrivateKeyPasswordKey]?.grantRead(documentRetrievalLambda);
-
-    return documentRetrievalLambda;
   }
 
   private setupBulkUrlSigningLambda(ownProps: {
